마이페이지-찜목록완료
예매율제외 구현완료

+++ src/main/java/com/star/cinema/myPage/service/MyPageServiceImpl.java
@@ -1,6 +1,8 @@
 package com.star.cinema.myPage.service;
 
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 import javax.servlet.http.HttpSession;
 
@@ -15,6 +17,9 @@ import com.star.cinema.customer.dto.QuestionDTO;
 import com.star.cinema.member.MemberCheck;
 import com.star.cinema.member.config.PageConfig;
 import com.star.cinema.member.dto.MemberDTO;
+import com.star.cinema.movie.dao.IGradeDAO;
+import com.star.cinema.movie.dto.GradeDTO;
+import com.star.cinema.movie.dto.MovieDTO;
 import com.star.cinema.myPage.dao.IMyPageDAO;
 
 @Service
@@ -22,6 +27,7 @@ public class MyPageServiceImpl implements IMyPageService {
 	final static Logger logger = LoggerFactory.getLogger(MyPageServiceImpl.class);
 	@Autowired HttpSession session;
 	@Autowired IMyPageDAO dao;
+	@Autowired IGradeDAO gradeDao;
 	@Autowired MemberCheck memChk;
 
 	@Override
@@ -42,7 +48,8 @@ public class MyPageServiceImpl implements IMyPageService {
 	@Override
 	public boolean modifyProc(String pw, String pwChk, String phone1, String phone2, String phone3, String zipcode,
 			String addr1, String addr2) {
-			
+		
+		if((MemberDTO)session.getAttribute("loginInfo") == null) return false;
 		if(!pw.equals(pwChk) || memChk.pwCheck(pwChk)== false || memChk.phoneCheck(phone2, phone3)== false ||
 				zipcode == "" || addr1 == "" || addr2 == "") {
 			return false;
@@ -62,7 +69,36 @@ public class MyPageServiceImpl implements IMyPageService {
 	@Override
 	public void deleteProc() {
 		MemberDTO dto = (MemberDTO)session.getAttribute("loginInfo");
+		if(dto == null) return;	
 		dao.deleteMember(dto);
+	}
+
+	@Override
+	public boolean myLikeProc(Model model) {
+		MemberDTO dto = (MemberDTO)session.getAttribute("loginInfo");
+		if(dto == null) return false;
+		ArrayList<GradeDTO> myGrade = gradeDao.selectMyLike(dto.getId());
+		Map<Integer,MovieDTO> myMovieInfo  = new HashMap<Integer,MovieDTO>();
+		Map<Integer,Double> myMovieGrade = new HashMap<Integer,Double>();
+		Map<Integer,String> recentReview = new HashMap<Integer,String>();
+		
+		if(!myGrade.isEmpty()) {
+			for(GradeDTO g : myGrade) {
+				MovieDTO movie = gradeDao.selectMovieInfo(g.getMovieListNum());
+				Double totalGrade = gradeDao.selectTotalGrade(g.getMovieListNum());
+				String recent = gradeDao.selectRecent(g.getMovieListNum());
+				myMovieInfo.put(g.getMovieListNum(), movie);
+				myMovieGrade.put(g.getMovieListNum(), totalGrade);
+				recentReview.put(g.getMovieListNum(),recent);
+			}	
+		}
+		
+		model.addAttribute("myGrade", myGrade);
+		model.addAttribute("myMovieInfo", myMovieInfo);
+		model.addAttribute("myMovieGrade", myMovieGrade);
+		model.addAttribute("recentReview", recentReview);
+		
+		return true;
 	}
